feat(suite-native): sheets blurred background overlay
Resolves #8576
