+++ suite-native/alerts/package.json
@@ -12,6 +12,7 @@
     "dependencies": {
         "@suite-common/icons": "workspace:*",
         "@suite-native/atoms": "workspace:*",
+        "@suite-native/screen-overlay": "workspace:*",
         "@trezor/styles": "workspace:*",
         "jotai": "1.9.1",
         "react": "18.2.0",
