+++ suite-native/atoms/package.json
@@ -15,6 +15,7 @@
         "@suite-common/icons": "workspace:*",
         "@suite-common/wallet-config": "workspace:*",
         "@suite-common/wallet-types": "workspace:*",
+        "@suite-native/screen-overlay": "workspace:*",
         "@suite-native/storage": "workspace:*",
         "@trezor/env-utils": "workspace:*",
         "@trezor/styles": "workspace:*",
