+++ suite-native/atoms/src/Sheet/BottomSheet.tsx
@@ -4,6 +4,7 @@ import Animated from 'react-native-reanimated';
 import { ScrollView, PanGestureHandler } from 'react-native-gesture-handler';
 
 import { prepareNativeStyle, useNativeStyles } from '@trezor/styles';
+import { BlurredScreenOverlay } from '@suite-native/screen-overlay';
 
 import { Box } from '../Box';
 import { BottomSheetContainer } from './BottomSheetContainer';
@@ -48,7 +49,6 @@ export const BottomSheet = ({
     const insets = useSafeAreaInsets();
     const [isCloseScrollEnabled, setIsCloseScrollEnabled] = useState(true);
     const {
-        animatedSheetWithOverlayStyle,
         animatedSheetWrapperStyle,
         closeSheetAnimated,
         openSheetAnimated,
@@ -79,9 +79,8 @@ export const BottomSheet = ({
 
     return (
         <BottomSheetContainer isVisible={isVisible} onClose={handleCloseSheet}>
-            <Animated.View
-                style={[animatedSheetWithOverlayStyle, applyStyle(sheetWithOverlayStyle)]}
-            >
+            <BlurredScreenOverlay />
+            <Box style={applyStyle(sheetWithOverlayStyle)}>
                 <PanGestureHandler
                     enabled={isCloseScrollEnabled}
                     ref={panGestureRef.current}
@@ -116,7 +115,7 @@ export const BottomSheet = ({
                         </ScrollView>
                     </Animated.View>
                 </PanGestureHandler>
-            </Animated.View>
+            </Box>
         </BottomSheetContainer>
     );
 };
