--- suite-native/screen-overlay/src/hooks/useScreenshot.ts
@@ -1,9 +1,0 @@
-import { useContext } from 'react';
-
-import { ScreenshotContext } from '../components/ScreenshotProvider';
-
-export const useScreenshot = () => {
-    const { screenshot } = useContext(ScreenshotContext);
-
-    return screenshot;
-};
