--- suite-native/screen-overlay/src/hooks/useTakeScreenshot.ts
@@ -1,9 +1,0 @@
-import { useContext } from 'react';
-
-import { ScreenshotContext } from '../components/ScreenshotProvider';
-
-export const useTakeScreenshot = () => {
-    const { takeScreenshot } = useContext(ScreenshotContext);
-
-    return takeScreenshot;
-};
