+++ suite-native/screen-overlay/src/index.ts
@@ -1,3 +1,3 @@
 export { BlurredScreenOverlay } from './components/BlurredScreenOverlay';
 export { ScreenshotProvider } from './components/ScreenshotProvider';
-export { useTakeScreenshot } from './hooks/useTakeScreenshot';
+export { ScreenshotCapturer } from './components/ScreenshotCapturer';
